Added auth to module endpoints.
Also added type warning suppression to database files that lacked it.

+++ backend/routers/module.py
@@ -4,16 +4,22 @@ from backend.models import ClientErrorResponse, ModuleDay, ModuleResponse
 from backend.dependencies import DBSession
 from backend.database import module as module_db
 
-
+from backend.exceptions import UnauthorizedException
+from fastapi import Depends
+from backend.auth import get_firebase_user_from_token
 
 router = APIRouter(prefix="/module", tags=["module"])
 
-@router.get("/{moduleID}/days",response_model=ModuleResponse, status_code=200,
+@router.get("/{moduleID}/days",
+            response_model=ModuleResponse, 
+            status_code=200,
             responses={
-                 404: {"model": ClientErrorResponse}
+                404: {"model": ClientErrorResponse}
              },
-            summary="Retrieve days for the given module.")
-def getModuleDays(moduleID:int,session:DBSession) -> ModuleResponse:
+            summary="Retrieve days for the given module. Must be an authenticated user.")
+def getModuleDays(moduleID:int,
+                  user: Annotated[dict, Depends(get_firebase_user_from_token)],
+                  session : DBSession) -> ModuleResponse:
     """Get the days of a module.
     
     Args:
@@ -38,8 +44,10 @@ def getModuleDays(moduleID:int,session:DBSession) -> ModuleResponse:
                  404: {"model": ClientErrorResponse},
                  409: {"model": ClientErrorResponse},
              },
-             summary="Create a new day within a module.")
-def create_new_day(moduleID: int, session: DBSession) -> ModuleDay:
+             summary="Create a new day within a module. Must be an authenticated owner of the module.")
+def create_new_day(moduleID: int, 
+                   user: Annotated[dict, Depends(get_firebase_user_from_token)],
+                   session: DBSession) -> ModuleDay:
     """ Create a new day within a module.
     
     Args:
@@ -52,6 +60,12 @@ def create_new_day(moduleID: int, session: DBSession) -> ModuleDay:
     Returns:
         ModuleDay: A response model containing the created classroom.
     """
+    
+    user_id = user["uid"]
+    teacher_id = module_db.get_teacher_by_module_id(moduleID, session)
+    if user_id != teacher_id:
+        raise UnauthorizedException("create day")
+    
     db_day = module_db.create_new_day(moduleID, session)
     return(ModuleDay(id=db_day.id, name = db_day.name)) # type: ignore
 
@@ -59,8 +73,10 @@ def create_new_day(moduleID: int, session: DBSession) -> ModuleDay:
 @router.delete("/{moduleID}",
                status_code=204,
                responses={404: {"model": ClientErrorResponse}},
-               summary="Delete a module.")
-def delete_classroom(moduleID: int, session: DBSession):
+               summary="Delete a module. Must be the authenticated owner of the module.")
+def delete_classroom(moduleID: int, 
+                     user: Annotated[dict, Depends(get_firebase_user_from_token)],
+                     session: DBSession):
     """Delete a module by its ID.
     Args:
         moduleID (int): The ID of the module to delete.
@@ -70,4 +86,9 @@ def delete_classroom(moduleID: int, session: DBSession):
     Returns:
         None
     """
+    user_id = user["uid"]
+    teacher_id = module_db.get_teacher_by_module_id(moduleID, session)
+    if user_id != teacher_id:
+        raise UnauthorizedException("create day")
+    
     module_db.delete_module(moduleID, session)
